requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
pywebview>=5.0.0; platform_system != "Windows"
